+++ src/background/services/openPayments.ts
@@ -1,19 +1,11 @@
 // cSpell:ignore keyid
-import type {
-  AccessToken,
-  AmountValue,
-  GrantDetails,
-  TabId,
-  WalletAmount,
-} from 'shared/types';
+import type { AmountValue, TabId } from 'shared/types';
 import {
   type AuthenticatedClient,
   createAuthenticatedClient,
   OpenPaymentsClientError,
 } from '@interledger/open-payments/dist/client';
 import {
-  isFinalizedGrant,
-  isPendingGrant,
   type IncomingPayment,
   type OutgoingPaymentWithSpentAmounts as OutgoingPayment,
   type WalletAddress,
@@ -22,8 +14,11 @@ import * as ed from '@noble/ed25519';
 import { type Request } from 'http-message-signatures';
 import { signMessage } from 'http-message-signatures/lib/httpbis';
 import { createContentDigestHeader } from 'httpbis-digest-headers';
-import type { Browser, Tabs } from 'webextension-polyfill';
-import { getExchangeRates, getRateOfPay, toAmount } from '../utils';
+import {
+  redirectToWelcomeScreen,
+  getExchangeRates,
+  getRateOfPay,
+} from '@/background/utils';
 import { KeyAutoAddService } from './keyAutoAdd';
 import { exportJWK, generateEd25519KeyPair } from '@/shared/crypto';
 import { bytesToHex } from '@noble/hashes/utils';
@@ -33,7 +28,6 @@ import {
   getWalletInformation,
   isErrorWithKey,
   sleep,
-  withResolvers,
   type ErrorWithKeyLike,
 } from '@/shared/helpers';
 import type {
@@ -49,18 +43,13 @@ import {
   OUTGOING_PAYMENT_POLLING_INTERVAL,
   OUTGOING_PAYMENT_POLLING_INITIAL_DELAY,
 } from '../config';
-import { OPEN_PAYMENTS_REDIRECT_URL } from '@/shared/defines';
 import type { Cradle } from '@/background/container';
+import { OutgoingPaymentGrantService } from './outgoingPaymentGrant';
+import { ErrorCode, GrantResult, InteractionIntent } from '@/shared/enums';
 
 interface KeyInformation {
   privateKey: string;
   keyId: string;
-}
-
-interface InteractionParams {
-  interactRef: string;
-  hash: string;
-  tabId: TabId;
 }
 
 export interface SignatureHeaders {
@@ -85,106 +74,52 @@ interface SignOptions {
   keyId: string;
 }
 
-interface VerifyInteractionHashParams {
-  clientNonce: string;
-  interactRef: string;
-  interactNonce: string;
-  hash: string;
-  authServer: string;
-}
-
-interface CreateOutgoingPaymentGrantParams {
-  clientNonce: string;
-  walletAddress: WalletAddress;
-  amount: WalletAmount;
-}
-
 interface CreateOutgoingPaymentParams {
   walletAddress: WalletAddress;
   incomingPaymentId: IncomingPayment['id'];
   amount: string;
 }
 
-type TabUpdateCallback = Parameters<Tabs.onUpdatedEvent['addListener']>[0];
-type TabRemovedCallback = Parameters<
-  Browser['tabs']['onRemoved']['addListener']
->[0];
-
-const enum ErrorCode {
-  CONTINUATION_FAILED = 'continuation_failed',
-  HASH_FAILED = 'hash_failed',
-  KEY_ADD_FAILED = 'key_add_failed',
-}
-
-const enum GrantResult {
-  GRANT_SUCCESS = 'grant_success',
-  GRANT_ERROR = 'grant_error',
-  KEY_ADD_SUCCESS = 'key_add_success',
-  KEY_ADD_ERROR = 'key_add_error',
-}
-
-const enum InteractionIntent {
-  CONNECT = 'connect',
-  RECONNECT = 'reconnect',
-  FUNDS = 'funds',
-  UPDATE_BUDGET = 'update_budget',
-}
-
 export class OpenPaymentsService {
   private browser: Cradle['browser'];
   private storage: Cradle['storage'];
-  private deduplicator: Cradle['deduplicator'];
-  private logger: Cradle['logger'];
+  private grantService: Cradle['grantService'];
   private appName: Cradle['appName'];
   private browserName: Cradle['browserName'];
   private t: Cradle['t'];
 
-  client?: AuthenticatedClient;
-
-  public switchGrant: OpenPaymentsService['_switchGrant'];
-
-  private token: AccessToken;
-  private grantDetails: GrantDetails | null;
-  /** Whether a grant has enough balance to make payments */
-  private isGrantUsable = { recurring: false, oneTime: false };
-
   constructor({
     browser,
     storage,
-    deduplicator,
-    logger,
-    t,
+    grantService,
     appName,
     browserName,
+    t,
   }: Cradle) {
     Object.assign(this, {
       browser,
       storage,
-      deduplicator,
-      logger,
-      t,
+      grantService,
       appName,
       browserName,
+      t,
     });
 
     void this.initialize();
-    this.switchGrant = this.deduplicator.dedupe(this._switchGrant.bind(this));
+  }
+
+  public client?: AuthenticatedClient;
+
+  public switchGrant(): OutgoingPaymentGrantService['switchGrant'] {
+    return this.grantService.switchGrant;
   }
 
   public isAnyGrantUsable() {
-    return this.isGrantUsable.recurring || this.isGrantUsable.oneTime;
+    return this.grantService.isAnyGrantUsable();
   }
 
-  private get grant() {
-    return this.grantDetails;
-  }
-
-  private set grant(grantDetails) {
-    this.logger.debug(`🤝🏻 Using grant: ${grantDetails?.type || null}`);
-    this.grantDetails = grantDetails;
-    this.token = grantDetails
-      ? grantDetails.accessToken
-      : { value: '', manageUrl: '' };
+  public async rotateToken() {
+    return this.grantService.rotateToken(this.client!);
   }
 
   private async initialize() {
@@ -196,15 +131,11 @@ export class OpenPaymentsService {
         'recurringGrant',
       ]);
 
-    this.isGrantUsable.recurring = !!recurringGrant;
-    this.isGrantUsable.oneTime = !!oneTimeGrant;
-
     if (
       connected === true &&
       walletAddress &&
       (recurringGrant || oneTimeGrant)
     ) {
-      this.grant = recurringGrant || oneTimeGrant!; // prefer recurring
       await this.initClient(walletAddress.id);
     }
   }
@@ -396,7 +327,8 @@ export class OpenPaymentsService {
     await this.initClient(walletAddress.id);
     this.setConnectState('connecting');
     try {
-      await this.completeGrant(
+      await this.grantService.createGrant(
+        this.client!,
         amount,
         walletAddress,
         recurring,
@@ -423,7 +355,8 @@ export class OpenPaymentsService {
         try {
           const tabId = await this.addPublicKeyToWallet(walletAddress);
           this.setConnectState('connecting');
-          await this.completeGrant(
+          await this.grantService.createGrant(
+            this.client!,
             amount,
             walletAddress,
             recurring,
@@ -457,7 +390,8 @@ export class OpenPaymentsService {
       'recurringGrant',
     ]);
 
-    await this.completeGrant(
+    await this.grantService.createGrant(
+      this.client!,
       amount,
       walletAddress!,
       recurring,
@@ -466,9 +400,15 @@ export class OpenPaymentsService {
 
     // cancel existing grants of same type, if any
     if (grants.oneTimeGrant && !recurring) {
-      await this.cancelGrant(grants.oneTimeGrant.continue);
+      await this.grantService.cancelGrant(
+        grants.oneTimeGrant.continue,
+        this.client!,
+      );
     } else if (grants.recurringGrant && recurring) {
-      await this.cancelGrant(grants.recurringGrant.continue);
+      await this.grantService.cancelGrant(
+        grants.recurringGrant.continue,
+        this.client!,
+      );
     }
 
     await this.storage.setState({ out_of_funds: false });
@@ -481,7 +421,8 @@ export class OpenPaymentsService {
       'recurringGrant',
     ]);
 
-    await this.completeGrant(
+    await this.grantService.createGrant(
+      this.client!,
       amount,
       walletAddress!,
       recurring,
@@ -492,7 +433,10 @@ export class OpenPaymentsService {
     // Note: Clear storage only if new grant type is not same as previous grant
     // type (as completeGrant already sets new grant state)
     if (existingGrants.oneTimeGrant) {
-      await this.cancelGrant(existingGrants.oneTimeGrant.continue);
+      await this.grantService.cancelGrant(
+        existingGrants.oneTimeGrant.continue,
+        this.client!,
+      );
       if (recurring) {
         this.storage.set({
           oneTimeGrant: null,
@@ -501,7 +445,10 @@ export class OpenPaymentsService {
       }
     }
     if (existingGrants.recurringGrant) {
-      await this.cancelGrant(existingGrants.recurringGrant.continue);
+      await this.grantService.cancelGrant(
+        existingGrants.recurringGrant.continue,
+        this.client!,
+      );
       if (!recurring) {
         this.storage.set({
           recurringGrant: null,
@@ -509,116 +456,6 @@ export class OpenPaymentsService {
         });
       }
     }
-  }
-
-  private async completeGrant(
-    amount: string,
-    walletAddress: WalletAddress,
-    recurring: boolean,
-    intent: InteractionIntent,
-    existingTabId?: TabId,
-  ): Promise<GrantDetails> {
-    const transformedAmount = toAmount({
-      value: amount,
-      recurring,
-      assetScale: walletAddress.assetScale,
-    });
-
-    const clientNonce = crypto.randomUUID();
-    const grant = await this.createOutgoingPaymentGrant({
-      clientNonce,
-      walletAddress,
-      amount: transformedAmount,
-    }).catch((err) => {
-      if (isInvalidClientError(err)) {
-        if (intent !== InteractionIntent.FUNDS) {
-          throw new ErrorWithKey('connectWallet_error_invalidClient');
-        }
-        const msg = this.t('connectWallet_error_invalidClient');
-        throw new Error(msg, { cause: err });
-      }
-      throw err;
-    });
-
-    const { interactRef, hash, tabId } = await this.getInteractionInfo(
-      grant.interact.redirect,
-      existingTabId,
-    );
-
-    await this.verifyInteractionHash({
-      clientNonce,
-      interactNonce: grant.interact.finish,
-      interactRef,
-      hash,
-      authServer: walletAddress.authServer,
-    }).catch(async (e) => {
-      await this.redirectToWelcomeScreen(
-        tabId,
-        GrantResult.GRANT_ERROR,
-        intent,
-        ErrorCode.HASH_FAILED,
-      );
-      throw e;
-    });
-
-    const continuation = await this.client!.grant.continue(
-      {
-        url: grant.continue.uri,
-        accessToken: grant.continue.access_token.value,
-      },
-      {
-        interact_ref: interactRef,
-      },
-    ).catch(async (e) => {
-      await this.redirectToWelcomeScreen(
-        tabId,
-        GrantResult.GRANT_ERROR,
-        intent,
-        ErrorCode.CONTINUATION_FAILED,
-      );
-      throw e;
-    });
-
-    if (!isFinalizedGrant(continuation)) {
-      throw new Error(
-        'Expected finalized grant. Received non-finalized grant.',
-      );
-    }
-
-    const grantDetails: GrantDetails = {
-      type: recurring ? 'recurring' : 'one-time',
-      amount: transformedAmount as Required<WalletAmount>,
-      accessToken: {
-        value: continuation.access_token.value,
-        manageUrl: continuation.access_token.manage,
-      },
-      continue: {
-        accessToken: continuation.continue.access_token.value,
-        url: continuation.continue.uri,
-      },
-    };
-
-    if (grantDetails.type === 'recurring') {
-      await this.storage.set({
-        recurringGrant: grantDetails,
-        recurringGrantSpentAmount: '0',
-      });
-      this.isGrantUsable.recurring = true;
-    } else {
-      await this.storage.set({
-        oneTimeGrant: grantDetails,
-        oneTimeGrantSpentAmount: '0',
-      });
-      this.isGrantUsable.oneTime = true;
-    }
-
-    this.grant = grantDetails;
-    await this.redirectToWelcomeScreen(
-      tabId,
-      GrantResult.GRANT_SUCCESS,
-      intent,
-    );
-    return grantDetails;
   }
 
   /**
@@ -644,7 +481,8 @@ export class OpenPaymentsService {
       const tabId = keyAutoAdd.tabId;
       const isTabClosed = error.key === 'connectWallet_error_tabClosed';
       if (tabId && !isTabClosed) {
-        await this.redirectToWelcomeScreen(
+        await redirectToWelcomeScreen(
+          this.browser,
           tabId,
           GrantResult.GRANT_ERROR,
           InteractionIntent.CONNECT,
@@ -676,145 +514,6 @@ export class OpenPaymentsService {
     });
   }
 
-  private async redirectToWelcomeScreen(
-    tabId: NonNullable<Tabs.Tab['id']>,
-    result: GrantResult,
-    intent: InteractionIntent,
-    errorCode?: ErrorCode,
-  ) {
-    const url = new URL(OPEN_PAYMENTS_REDIRECT_URL);
-    url.searchParams.set('result', result);
-    url.searchParams.set('intent', intent);
-    if (errorCode) url.searchParams.set('errorCode', errorCode);
-    await this.browser.tabs.update(tabId, {
-      url: url.toString(),
-    });
-  }
-
-  private async createOutgoingPaymentGrant({
-    amount,
-    walletAddress,
-    clientNonce,
-  }: CreateOutgoingPaymentGrantParams) {
-    const grant = await this.client!.grant.request(
-      {
-        url: walletAddress.authServer,
-      },
-      {
-        access_token: {
-          access: [
-            {
-              type: 'quote',
-              actions: ['create'],
-            },
-            {
-              type: 'outgoing-payment',
-              actions: ['create', 'read'],
-              identifier: walletAddress.id,
-              limits: {
-                debitAmount: {
-                  value: amount.value,
-                  assetScale: walletAddress.assetScale,
-                  assetCode: walletAddress.assetCode,
-                },
-                interval: amount.interval,
-              },
-            },
-          ],
-        },
-        interact: {
-          start: ['redirect'],
-          finish: {
-            method: 'redirect',
-            uri: OPEN_PAYMENTS_REDIRECT_URL,
-            nonce: clientNonce,
-          },
-        },
-      },
-    );
-
-    if (!isPendingGrant(grant)) {
-      throw new Error(
-        'Expected interactive grant. Received non-pending grant.',
-      );
-    }
-
-    return grant;
-  }
-
-  private async verifyInteractionHash({
-    clientNonce,
-    interactRef,
-    interactNonce,
-    hash,
-    authServer,
-  }: VerifyInteractionHashParams): Promise<void> {
-    const grantEndpoint = new URL(authServer).origin + '/';
-    const data = new TextEncoder().encode(
-      `${clientNonce}\n${interactNonce}\n${interactRef}\n${grantEndpoint}`,
-    );
-
-    const digest = await crypto.subtle.digest('SHA-256', data);
-    const calculatedHash = btoa(
-      String.fromCharCode.apply(null, new Uint8Array(digest)),
-    );
-    if (calculatedHash !== hash) throw new Error('Invalid interaction hash');
-  }
-
-  private async getInteractionInfo(
-    url: string,
-    existingTabId?: TabId,
-  ): Promise<InteractionParams> {
-    const { resolve, reject, promise } = withResolvers<InteractionParams>();
-
-    const tab = existingTabId
-      ? await this.browser.tabs.update(existingTabId, { url })
-      : await this.browser.tabs.create({ url });
-    if (!tab.id) {
-      reject(new Error('Could not create/update tab'));
-      return promise;
-    }
-
-    const tabCloseListener: TabRemovedCallback = (tabId) => {
-      if (tabId !== tab.id) return;
-
-      this.browser.tabs.onRemoved.removeListener(tabCloseListener);
-      reject(new ErrorWithKey('connectWallet_error_tabClosed'));
-    };
-
-    const getInteractionInfo: TabUpdateCallback = async (tabId, changeInfo) => {
-      if (tabId !== tab.id) return;
-      try {
-        const tabUrl = new URL(changeInfo.url || '');
-        const interactRef = tabUrl.searchParams.get('interact_ref');
-        const hash = tabUrl.searchParams.get('hash');
-        const result = tabUrl.searchParams.get('result');
-
-        if (
-          (interactRef && hash) ||
-          result === 'grant_rejected' ||
-          result === 'grant_invalid'
-        ) {
-          this.browser.tabs.onUpdated.removeListener(getInteractionInfo);
-          this.browser.tabs.onRemoved.removeListener(tabCloseListener);
-        }
-
-        if (interactRef && hash) {
-          resolve({ interactRef, hash, tabId });
-        } else if (result === 'grant_rejected') {
-          reject(new ErrorWithKey('connectWallet_error_grantRejected'));
-        }
-      } catch {
-        /* do nothing */
-      }
-    };
-
-    this.browser.tabs.onRemoved.addListener(tabCloseListener);
-    this.browser.tabs.onUpdated.addListener(getInteractionInfo);
-
-    return promise;
-  }
-
   async disconnectWallet() {
     const { recurringGrant, oneTimeGrant } = await this.storage.get([
       'recurringGrant',
@@ -824,27 +523,17 @@ export class OpenPaymentsService {
       return;
     }
     if (recurringGrant) {
-      await this.cancelGrant(recurringGrant.continue);
-      this.isGrantUsable.recurring = false;
+      await this.grantService.cancelGrant(
+        recurringGrant.continue,
+        this.client!,
+      );
+      this.grantService.disableRecurringGrant();
     }
     if (oneTimeGrant) {
-      await this.cancelGrant(oneTimeGrant.continue);
-      this.isGrantUsable.oneTime = false;
+      await this.grantService.cancelGrant(oneTimeGrant.continue, this.client!);
+      this.grantService.disableOneTimeGrant();
     }
     await this.storage.clear();
-    this.grant = null;
-  }
-
-  private async cancelGrant(grantContinuation: GrantDetails['continue']) {
-    try {
-      await this.client!.grant.cancel(grantContinuation);
-    } catch (error) {
-      if (isInvalidClientError(error) || isInvalidContinuationError(error)) {
-        // key already removed from wallet
-        return;
-      }
-      throw error;
-    }
   }
 
   async generateKeys() {
@@ -868,7 +557,7 @@ export class OpenPaymentsService {
   }: CreateOutgoingPaymentParams): Promise<OutgoingPayment> {
     const outgoingPayment = (await this.client!.outgoingPayment.create(
       {
-        accessToken: this.token.value,
+        accessToken: this.grantService.accessToken(),
         url: walletAddress.resourceServer,
       },
       {
@@ -887,7 +576,7 @@ export class OpenPaymentsService {
 
     if (outgoingPayment.grantSpentDebitAmount) {
       this.storage.updateSpentAmount(
-        this.grant!.type,
+        this.grantService.grantType(),
         outgoingPayment.grantSpentDebitAmount.value,
       );
     }
@@ -911,7 +600,7 @@ export class OpenPaymentsService {
         signal?.throwIfAborted();
         const outgoingPayment = await this.client!.outgoingPayment.get({
           url: outgoingPaymentId,
-          accessToken: this.token.value,
+          accessToken: this.grantService.accessToken(),
         });
         yield outgoingPayment;
         if (
@@ -961,7 +650,7 @@ export class OpenPaymentsService {
     await this.client!.quote.create(
       {
         url: sender.resourceServer,
-        accessToken: this.token.value,
+        accessToken: this.grantService.accessToken(),
       },
       {
         method: 'ilp',
@@ -1019,7 +708,8 @@ export class OpenPaymentsService {
         await this.validateReconnect();
 
         tabId ??= await this.ensureTabExists();
-        await this.redirectToWelcomeScreen(
+        await redirectToWelcomeScreen(
+          this.browser,
           tabId,
           GrantResult.KEY_ADD_SUCCESS,
           InteractionIntent.RECONNECT,
@@ -1027,7 +717,8 @@ export class OpenPaymentsService {
       } catch (error) {
         const isTabClosed = error.key === 'connectWallet_error_tabClosed';
         if (tabId && !isTabClosed) {
-          await this.redirectToWelcomeScreen(
+          await redirectToWelcomeScreen(
+            this.browser,
             tabId,
             GrantResult.KEY_ADD_ERROR,
             InteractionIntent.RECONNECT,
@@ -1047,58 +738,6 @@ export class OpenPaymentsService {
       throw new Error('Could not create tab');
     }
     return tab.id;
-  }
-
-  /**
-   * Switches to the next grant that can be used.
-   * @returns the type of grant that should be used now, or null if no grant can
-   * be used.
-   */
-  private async _switchGrant(): Promise<GrantDetails['type'] | null> {
-    if (!this.isAnyGrantUsable()) {
-      return null;
-    }
-    this.logger.debug('Switching from grant', this.grant?.type);
-    const { oneTimeGrant, recurringGrant } = await this.storage.get([
-      'oneTimeGrant',
-      'recurringGrant',
-    ]);
-    if (this.grant?.type === 'recurring') {
-      this.isGrantUsable.recurring = false;
-      if (oneTimeGrant) {
-        this.grant = oneTimeGrant;
-        return 'one-time';
-      }
-    } else if (this.grant?.type === 'one-time') {
-      this.isGrantUsable.oneTime = false;
-      if (recurringGrant) {
-        this.grant = recurringGrant;
-        return 'recurring';
-      }
-    }
-    return null;
-  }
-
-  async rotateToken() {
-    if (!this.grant) {
-      throw new Error('No grant to rotate token for');
-    }
-    const rotate = this.deduplicator.dedupe(this.client!.token.rotate);
-    const newToken = await rotate({
-      url: this.token.manageUrl,
-      accessToken: this.token.value,
-    });
-    const accessToken: AccessToken = {
-      value: newToken.access_token.value,
-      manageUrl: newToken.access_token.manage,
-    };
-    if (this.grant.type === 'recurring') {
-      this.storage.set({ recurringGrant: { ...this.grant, accessToken } });
-    } else {
-      this.storage.set({ oneTimeGrant: { ...this.grant, accessToken } });
-    }
-    this.grant = { ...this.grant, accessToken };
-    return newToken;
   }
 }
